fix: remove shebang flag and use fs.readJSONSync to read JSON file (#58)
Importing JSON modules causing a warning to appear regarding that
it is experimental. In order to suppress that warning a flag was
added on index.js but it is not portable and was causing zem to
freeze CI steps on Github Actions. There is no other practical way
to supress warning messages, hence this commit replaces JSON imports
with fs method to address this issue.